ShowlistPreview.swift cleanup
