+++ ShotlistDevChallenge/ContentView.swift
@@ -12,7 +12,7 @@ struct ContentView: View {
   @State var onClick: OnClickHandler = { }
 
   var body: some View {
-      ShootPreview()
+      ShotlistPreview()
     }
   }
